+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 
 - Run `ecomp` tests
+- Test results are reset after window reload
 
 ## [0.4.0] - 2022-09-10